+++ release_notes.md
@@ -5,6 +5,7 @@
 -   APP-4306 - [API] Added support for custom association pagination
 -   APP-4381 - [Logging] Fix API logging issues
 -   APP-4383 - [API] Update for changes in the TC V3 API
+-   APP-4400 - [Input] Add support for KeyValue input type with None value
 
 ## 4.0.3